pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "botted-library"
version = "1.0.0"
authors = [
    {name = "Botted Library Team", email = "contact@botted-library.com"},
]
description = "Human-like AI workers that can use any tool to accomplish tasks - web search, coding, document creation, and more"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
keywords = ["ai", "automation", "browser", "selenium", "llm", "workers", "intelligent", "agents", "human-like", "research", "coding"]
dependencies = [
    "selenium>=4.15.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "webdriver-manager>=4.0.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
openai = ["openai>=1.3.0"]
anthropic = ["anthropic>=0.7.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://botted-library.com"
Documentation = "https://github.com/botted-library/botted-library#readme"
Repository = "https://github.com/botted-library/botted-library"
"Bug Reports" = "https://github.com/botted-library/botted-library/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["botted_library*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
botted_library = ["*.md", "*.txt", "*.json"]